requirements.txt
matrix-nio[e2e]
markdown
pyephem
astronomy-engine
geopy
timezonefinder
anthropic
